Add files via upload
This file contains code of minimax algorithm which is used to get the solution for multiplayer game where players play consecutively against each other and assumed that they play steps optimally. That is they chose most optimal step for their turn. And using backtracking and this we get results of such games